move shim related functions into separate namespace
